Implementa exibição dos dados do usuário retornados pelo backend
