+++ src/pages/PerfilCoordenador/PerfilCoordenador.tsx
@@ -1,17 +1,45 @@
 import "./styles.css";
 
 import { useState } from "react";
-import { MdGroupAdd, MdLogout } from "react-icons/md";
+import { MdGroupAdd, MdLogout, MdOutlineLibraryBooks } from "react-icons/md"; // Importando ícones
+import { useNavigate } from "react-router-dom";
 
+import { Professor } from "@/models/User";
 import { useUserQueries } from "@/queries/user";
 
-import Solicitacoes from "../../components/Solicitacoes/Solicitacoes";
-
-function PerfilOrientador() {
-  const { signOut } = useUserQueries();
+function PerfilCoordenador() {
   const logoPgcop = "/assets/logoPgcop.png";
 
+  const navigate = useNavigate();
+
+  const { signOut, useGetUser } = useUserQueries();
+
+  const { data: userData } = useGetUser();
+
+  const user = userData as Professor;
+
   const alunosData = [
+    {
+      id: 1,
+      nome: "João Silva",
+      matricula: "2022001",
+      titulacao: "Mestrado",
+      datafinal: "03/05/2024",
+    },
+    {
+      id: 1,
+      nome: "João Silva",
+      matricula: "2022001",
+      titulacao: "Mestrado",
+      datafinal: "03/05/2024",
+    },
+    {
+      id: 1,
+      nome: "João Silva",
+      matricula: "2022001",
+      titulacao: "Mestrado",
+      datafinal: "03/05/2024",
+    },
     {
       id: 1,
       nome: "João Silva",
@@ -24,7 +52,7 @@ function PerfilOrientador() {
       nome: "Tauane Souza",
       matricula: "2022002",
       titulacao: "Doutorado",
-      datafinal: "8/05/2027",
+      datafinal: "18/05/2027",
     },
     {
       id: 3,
@@ -63,34 +91,9 @@ function PerfilOrientador() {
     },
   ];
 
-  const [solicitacoes, setSolicitacoes] = useState([
-    {
-      id: 1,
-      nome: "Natalia  Santos Santos Santos",
-      matricula: "2022001",
-      titulacao: "Mestrado",
-      datafinal: "03/05/2024",
-    },
-    {
-      id: 2,
-      nome: "Claudio Souza",
-      matricula: "2022002",
-      titulacao: "Doutorado",
-      datafinal: "8/05/2027",
-    },
-    {
-      id: 3,
-      nome: "Vinicius Alves",
-      matricula: "2022003",
-      titulacao: "Mestrado",
-      datafinal: "15/05/2024",
-    },
-  ]);
-
   const [alunos, setAlunos] = useState(alunosData);
   const [showModal, setShowModal] = useState(false);
   const [selectedAluno, setSelectedAluno] = useState(null);
-  const [showSolicitacoes, setShowSolicitacoes] = useState(false);
 
   const handleDoubleClick = (matricula) => {
     const aluno = alunos.find((aluno) => aluno.matricula === matricula);
@@ -107,28 +110,7 @@ function PerfilOrientador() {
     setShowModal(false);
   };
 
-  const handleSolicitacoesClick = () => {
-    setShowSolicitacoes(!showSolicitacoes);
-  };
-
-  const handleAcceptRequest = (id) => {
-    const solicitationToAccept = solicitacoes.find(
-      (solicitacao) => solicitacao.id === id,
-    );
-    setAlunos([...alunos, solicitationToAccept]);
-    const updatedSolicitacoes = solicitacoes.filter(
-      (solicitacao) => solicitacao.id !== id,
-    );
-    setSolicitacoes(updatedSolicitacoes);
-  };
-
-  const handleRemoveRequest = (id) => {
-    const updatedSolicitacoes = solicitacoes.filter(
-      (solicitacao) => solicitacao.id !== id,
-    );
-    setSolicitacoes(updatedSolicitacoes);
-  };
-
+  // Separa os alunos baseados na titulação
   const alunosMestrado = alunos.filter(
     (aluno) => aluno.titulacao === "Mestrado",
   );
@@ -138,45 +120,40 @@ function PerfilOrientador() {
 
   return (
     <div className="contain">
-      <div className="containerOrientador">
-        <img src={logoPgcop} />
+      <div className="containerCoordenador">
+        {/* Logo*/}
+        <img src={logoPgcop} alt="Logo" />
+        {/* Informações do perfil */}
         <div
-          className="infoOrientador"
-          style={{ justifyContent: "space-between" }}
+          className="infoCoordenador"
+          style={{ justifyContent: "space-between", marginRight: "30px" }}
         >
           <div>
-            <h2>Augusto Carlos Santos</h2>
+            <h2>{user?.nome}</h2>
             <h3>Orientandos: {alunos.length}</h3>
           </div>
-          <div className="botoesToolbar">
-            <div style={{ position: "relative" }}>
+          {/* Botões Toolbar */}
+          <div>
+            <div className="botoesToolbar">
               <MdGroupAdd
-                onClick={handleSolicitacoesClick}
-                style={{
-                  marginRight: "40px",
-                  cursor: "pointer",
-                  color: solicitacoes.length > 0 ? "red" : "inherit",
-                }}
+                onClick={() => navigate("/perfil-coordenador/solicitacoes")}
+                style={{ cursor: "pointer", marginRight: "42px" }}
                 size={35}
                 title="Solicitações"
               />
-              {showSolicitacoes && (
-                <div
-                  className="solicitacoesContainer"
-                  style={{ position: "absolute", top: "-50px" }}
-                >
-                  <Solicitacoes
-                    solicitacoes={solicitacoes}
-                    handleAcceptRequest={handleAcceptRequest}
-                    handleRemoveRequest={handleRemoveRequest}
-                  />
-                </div>
-              )}
+            </div>
+            <div>
+              <MdOutlineLibraryBooks
+                onClick={() => navigate("/tarefas")}
+                style={{ cursor: "pointer", marginRight: "45px" }}
+                size={35}
+                title="Tarefas"
+              />
             </div>
             <div>
               <MdLogout
                 onClick={signOut}
-                style={{ marginRight: "40px", cursor: "pointer" }}
+                style={{ cursor: "pointer", marginRight: "30px" }}
                 size={35}
                 title="Sair"
               />
@@ -188,8 +165,8 @@ function PerfilOrientador() {
       <h2 style={{ textAlign: "center", marginTop: "40px" }}>
         Lista de Orientandos
       </h2>
-
-      <div className="containerOrientadorOrientandos">
+      {/* Container de Alunos Orientados - Mestrado */}
+      <div className="containerOrientandosCoordenador">
         <h3 style={{ textAlign: "center", marginBottom: "10px" }}>
           Alunos de Mestrado
         </h3>
@@ -242,8 +219,9 @@ function PerfilOrientador() {
         </ul>
       </div>
 
+      {/* Container de Alunos Orientados - Doutorado */}
       <div
-        className="containerOrientadorOrientandos"
+        className="containerOrientandosCoordenador"
         style={{ marginTop: "30px" }}
       >
         <h3 style={{ textAlign: "center", marginBottom: "10px" }}>
@@ -298,13 +276,14 @@ function PerfilOrientador() {
         </ul>
       </div>
 
+      {/* Modal de confirmação */}
       {showModal && (
         <div className="confirmationBox">
           <div
             style={{
               backgroundColor: "#fff",
-              padding: "10px",
-              borderRadius: "8px",
+              padding: "20px",
+              borderRadius: "10px",
               width: "300px",
               textAlign: "center",
             }}
@@ -333,4 +312,4 @@ function PerfilOrientador() {
   );
 }
 
-export default PerfilOrientador;
+export default PerfilCoordenador;
